+++ src/water_masses/__init__.py
@@ -1,14 +1,37 @@
 # -*- coding: utf-8 -*-
-from . import tracmass, spgsi, example, processing, version, filter_month
+from . import (
+    version,
+    # modules
+    constants,
+    spgsi,
+    processing,
+    filter_month,
+    time_series,
+    filtering,
+    transform,
+    # submodules
+    tracmass,
+    origin,
+)
+from .version import pkg_version as __version__
 from .tracmass import seeding
-
+from .origin import pca
 
 __all__ = [
-    "tracmass",
-    "seeding",
-    "example",
-    "processing",
     "version",
+    "__version__",
+    # modules
+    "constants",
+    "processing",
     "filter_month",
     "spgsi",
+    "time_series",
+    "filtering",
+    "transform",
+    # submodule
+    "tracmass",
+    "seeding",
+    # submodule
+    "origin",
+    "pca",
 ]