+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 We follow [Calendar Versions](https://calver.org/).
 
+## Version 2021.3
+
+- Add PC Analysis
+- Add time series tools
+- Add apply function for lowpass and bandstop time-filter to field
+- Add function to open tracmass files
+
 ## Version 2021.2
 
 - Experimental designs